fix: Convert CSS to styled-componenet
