+++ vite-project/package.json
@@ -13,6 +13,7 @@
     "classnames": "^2.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-kakao-link": "^0.1.2",
     "react-router-dom": "^6.20.1",
     "styled-components": "^6.1.1"
   },
