+++ vite-project/src/components/FolderHeader.jsx
@@ -1,17 +1,144 @@
-import * as S from "./styled";
+import { useState } from "react";
+import styled from "styled-components";
+import FolderList from "./common/FolderList";
+import Button from "./common/Button";
+import Modal from "./common/Modal";
 import linkIcon from "../../../images/folder/link.svg";
 
-function FolderHeader() {
+const flex = `
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+const StyledFolderHeader = styled.header`
+  padding: 6rem 0 9rem;
+  background-color: #f0f6ff;
+
+  form {
+    width: 80rem;
+    display: flex;
+    gap: 1.2rem;
+    align-items: center;
+    background: #fff;
+    border-radius: 1.5rem;
+    border: 1px solid #6d6afe;
+    padding: 1.6rem 2rem;
+    margin: 0 auto;
+  }
+
+  .link-icon {
+    width: 2rem;
+    height: 2rem;
+    align-items: center;
+    border: none;
+    background-color: transparent;
+    ${flex}
+  }
+
+  input {
+    width: 100%;
+    border: none;
+    outline: none;
+    background: #fff;
+
+    &::placeholder {
+      color: #9fa6b2;
+      font-size: 1.6rem;
+      font-weight: 400;
+      line-height: 2.4rem;
+    }
+  }
+
+  .link-cta {
+    width: 8rem;
+    border: none;
+    border-radius: 0.8rem;
+    background-image: linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%);
+    color: #f5f5f5;
+    font-size: 1.4rem;
+    font-weight: 600;
+    padding: 1rem 1.6rem;
+    align-items: center;
+    white-space: nowrap;
+    ${flex}
+  }
+
+  @media (max-width: 1124px) {
+    padding: 6rem 3.25rem 9rem;
+
+    form {
+      width: 70.4rem;
+    }
+  }
+
+  @media (max-width: 767px) {
+    padding: 2.4rem 3.2rem 4rem;
+
+    form {
+      width: 32.5rem;
+      padding: 0.8rem 1rem;
+    }
+
+    .link-icon {
+      width: 1.6rem;
+      height: 1.6rem;
+
+      img {
+        width: 1.6rem;
+        height: 1.6rem;
+      }
+    }
+  }
+`;
+
+function FolderHeader({ folderList }) {
+  const [isAddModalOpen, setIsAddModalOpen] = useState(false);
+  const [inputUrl, setInputUrl] = useState("");
+
+  const handleAddModal = (e) => {
+    e.preventDefault();
+    setIsAddModalOpen(true);
+  };
+
+  const handleCloseModal = () => {
+    setIsAddModalOpen(false);
+  };
+
+  const handleInputUrlChange = (e) => {
+    setInputUrl(e.target.value);
+  };
+
   return (
-    <S.FolderHeader>
+    <StyledFolderHeader>
       <form>
         <button className="link-icon">
           <img src={linkIcon} alt="첨부 아이콘" />
         </button>
-        <input placeholder="링크를 추가해 보세요." />
-        <button className="link-cta">추가하기</button>
+        <input
+          placeholder="링크를 추가해 보세요."
+          onChange={handleInputUrlChange}
+        />
+        <button className="link-cta" onClick={handleAddModal}>
+          추가하기
+        </button>
       </form>
-    </S.FolderHeader>
+
+      {isAddModalOpen && (
+        <Modal
+          modalTitle="폴더에 추가"
+          subTitle={inputUrl}
+          onClose={handleCloseModal}
+        >
+          <div className="modal-content">
+            <FolderList folderList={folderList} />
+            <Button variant="default" size="lg">
+              추가하기
+            </Button>
+          </div>
+        </Modal>
+      )}
+    </StyledFolderHeader>
   );
 }
