--- vite-project/src/components/FolderMainCard.css
@@ -1,29 +1,0 @@
-.kebab-menu {
-  position: absolute;
-  z-index: 2;
-
-  list-style: none;
-  box-shadow: 0px 2px 8px 0px rgba(51, 50, 54, 0.1);
-}
-
-.hide {
-  display: none;
-  opacity: 0;
-  visibility: hidden;
-}
-
-.kebab-menu button {
-  display: block;
-  width: 100px;
-  padding: 7px 0;
-  font-size: 14px;
-  font-weight: 400;
-  background-color: #fff;
-  border: none;
-  outline: none;
-}
-
-.kebab-menu li:hover button {
-  color: var(--primary);
-  background-color: #e7effb;
-}
