--- vite-project/src/components/common/FolderList.css
@@ -1,37 +1,0 @@
-.add-folder-list {
-  width: 100%;
-}
-
-.add-folder-list-item {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  padding: 8px;
-  border-radius: 8px;
-  background-color: #fff;
-  cursor: pointer;
-}
-
-.add-folder-list-item:not(:last-child) {
-  margin-bottom: 4px;
-}
-
-.add-folder-list-item:hover {
-  background-color: #f0f6ff;
-}
-
-.add-folder-list-item:hover p {
-  color: #6d6afe;
-}
-
-.add-folder-list-item p {
-  font-size: 16px;
-  line-height: 24px;
-  color: #373740;
-}
-
-.add-folder-list-item span {
-  font-size: 14px;
-  color: #9fa6b2;
-  margin-left: 8px;
-}
