--- vite-project/src/components/common/button.css
@@ -1,34 +1,0 @@
-.btn-base {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  font-size: 16px;
-  font-weight: 600;
-  color: #f5f5f5;
-  padding: 16px 0;
-  border-radius: 8px;
-  outline: none;
-  border: none;
-}
-
-.btn-size-lg {
-  width: 280px;
-}
-
-.btn-size-md {
-}
-
-.btn-size-sm {
-}
-
-.btn-default {
-  background: var(
-    --gra-purpleblue-to-skyblue,
-    linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%)
-  );
-}
-
-.btn-remove {
-  background-color: var(--red);
-}
