--- vite-project/src/components/common/input.css
@@ -1,15 +1,0 @@
-.input {
-  width: 100%;
-  padding: 18px 15px;
-  font-size: 16px;
-  color: var(--gray100);
-  border-radius: 8px;
-  border: 1px solid var(--gray20);
-  background: var(--white);
-  outline: none;
-}
-
-.input:hover,
-.input:focus {
-  border: 1px solid var(--primary);
-}
