--- vite-project/src/components/common/modal.css
@@ -1,50 +1,0 @@
-.modal {
-  position: relative;
-
-  padding: 32px 40px;
-  border: 1px solid #ccd5e3;
-  border-radius: 15px;
-  background-color: var(--white);
-}
-
-.modal-header {
-  text-align: center;
-  margin-bottom: 24px;
-}
-
-.modal-header-title {
-  font-size: 20px;
-  font-weight: 700;
-  color: var(--gray100);
-}
-
-.modal-btn-close {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 24px;
-  height: 24px;
-  border: none;
-  outline: none;
-  border-radius: 50%;
-  background-color: var(--gray10);
-
-  position: absolute;
-  top: 15px;
-  right: 15px;
-}
-
-.modal-header-sub-title {
-  color: var(--gray60, #9fa6b2);
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 22px;
-  margin-top: 8px;
-}
-
-.modal-content {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
