--- vite-project/src/components/common/shareLink.css
@@ -1,24 +1,0 @@
-.share-list {
-  display: flex;
-  gap: 32px;
-  list-style: none;
-}
-
-.share-list-item {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.share-list-item img {
-  width: 42px;
-  height: 42px;
-}
-
-.share-list-item p {
-  font-size: 13px;
-  color: #373740;
-  margin-top: 10px;
-  white-space: nowrap;
-}
