+++ vite-project/src/stores/constants.js
@@ -1,0 +1,31 @@
+import facebookIcon from "../../../images/landing/icons_facebook.svg";
+import twitterIcon from "../../../images/landing/icons_twitter.svg";
+import youtubeIcon from "../../../images/landing/icons_youtube.svg";
+import instagramIcon from "../../../images/landing/icons_instagram.svg";
+
+export const SOCIAL_LIST = [
+  {
+    id: 1,
+    name: "페이스북",
+    icon: facebookIcon,
+    link: "https://www.facebook.com/?locale=ko_KR",
+  },
+  {
+    id: 2,
+    name: "트위터",
+    icon: twitterIcon,
+    link: "https://twitter.com/?lang=ko",
+  },
+  {
+    id: 3,
+    name: "유튜브",
+    icon: youtubeIcon,
+    link: "https://www.youtube.com/",
+  },
+  {
+    id: 4,
+    name: "인스타그램",
+    icon: instagramIcon,
+    link: "https://www.instagram.com/",
+  },
+];
